requirements.txt
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
Flask==2.0.1
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
MarkupSafe==2.0.1
numpy==1.20.3
pandas==1.2.4
psycopg2==2.8.6
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
scikit-learn==0.24.2
scipy==1.6.3
six==1.16.0
sklearn==0.0
threadpoolctl==2.1.0
urllib3==1.26.5
waitress==2.0.0
Werkzeug==2.0.1
